pyproject.toml
[tool.poetry]
name = "instagram-server-django"
version = "0.1.0"
description = ""
authors = ["INSEOP KIM <inseop0813@gmail.com>"]
readme = "README.md"
packages = [{include = "instagram_server_django"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pyjwt = "^2.6.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
